requirements.txt
streamlit
pandas
faiss-cpu
numpy
openai
